Working paths to posts(old id.tsx changed to .md)
